reworked arm nodes variable types

+++ ROS Nodes/arm_base.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 import rospy
-from std_msgs.msg import Int16
+from std_msgs.msg import Int32
 
 def callback(data):
     rospy.loginfo(data.data)
@@ -8,7 +8,7 @@ def callback(data):
 def arm_base():
 
     rospy.init_node('arm_base')
-    rospy.Subscriber('base', Int16, callback)
+    rospy.Subscriber('base', Int32, callback)
     rospy.spin()
 
 if __name__ == '__main__':
